refactor(data-structures): move serialization conversion error to error.rs
